source1/requirements.txt
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
tablib==2.0.0
whitenoise==5.1.0
